[REFACTOR] Move add instruction article flow into new component
